qcacmn: Changes to store stats in 64bit data type
Changed the data type to store stats to 64bit
from 32bit to support larger number for stats
counter

CRs-Fixed: 2024225
Acked-by: Chaithanya Garrepalli <cgarre@codeaurora.org>

+++ dp/inc/cdp_txrx_stats_struct.h
@@ -52,13 +52,13 @@
 
 #ifndef BIG_ENDIAN_HOST
 typedef struct {
-	u_int32_t pkts;
-	u_int32_t bytes;
+	uint64_t pkts;
+	uint64_t bytes;
 } ol_txrx_stats_elem;
 #else
 struct ol_txrx_elem_t {
-	u_int32_t pkts;
-	u_int32_t bytes;
+	uint64_t pkts;
+	uint64_t bytes;
 };
 typedef struct ol_txrx_elem_t ol_txrx_stats_elem;
 #endif
